src/test/java/com/leetcode/algorithm/medium/inordersuccessorinbst/SolutionTest.java
package com.leetcode.algorithm.medium.inordersuccessorinbst;

import com.leetcode.algorithm.common.TreeNode;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNull;

public class SolutionTest {
  @Test
  public void testCase1() {
    Solution solution = new Solution();
    TreeNode root = new TreeNode(2);
    root.left = new TreeNode(1);
    root.right = new TreeNode(3);

    TreeNode result = solution.inorderSuccessor(root, root.left);
    assertEquals(root, result);
  }

  @Test
  public void testCase2() {
    Solution solution = new Solution();
    TreeNode root = new TreeNode(5);
    root.left = new TreeNode(3);
    root.left.left = new TreeNode(2);
    root.left.right = new TreeNode(4);
    root.left.left.left = new TreeNode(1);
    root.right = new TreeNode(6);

    TreeNode result = solution.inorderSuccessor(root, root.right);
    assertNull(result);
  }
}
